feat: add javascript
I already had some of the code written elsewhere, i just needed the knowledge to put it into a nice page without too much hassle. I'll add the reference to calculate the current hour in the README.md so if the calculations don't make sense, you can see how i got them from.
The script basically replaces the "00" in the html and replaces them with the correct current metric. It autoupdates everysecond so it will run to eternity!!